requirements.txt
ipywidgets>=8.0.0
yt-dlp>=2023.12.30
whisperx>=3.0.0
torch>=2.0.0
google-generativeai>=0.3.2
python-dotenv>=1.0.0
tenacity>=8.0.0
urllib3<2.0.0  # To avoid OpenSSL warning
pydub>=0.25.1
ratelimit>=2.2.1 